Only show the Tweet button if the user can edit the played game
A user can edit the played game if they are logged in and their
current gaming group is the same as the gaming group from the played
game.

This change makes the Tweet button also require the above conditions,
so the user must have the played game's group as their current active
gaming group.